Move toggleBookmark function from utils to bookmarkModule
